requirements.txt
bs4
praw
pyTelegramBotAPI
python-decouple
requests
selenium
